refactor: apply reviews for update modal
- Fix inverted installOnQuit condition in update-downloaded handler
- Replace stale line-number comment with descriptive text
- Separate useUpdateStore into updateStore.ts
- Rename types/setting.ts to types/update.ts for clarity
- Subscribe to UPDATE_NOT_AVAILABLE and show toast notification
- Show toast on UPDATE_ERROR with error message
- Display releaseNotes in UpdateModal with scrollable container
- Disable action buttons while downloading to prevent duplicate actions
