+++ src/renderer/src/components/UpdateModal.vue
@@ -7,7 +7,7 @@
     DialogHeader,
     DialogTitle,
   } from '@/components/ui/dialog';
-  import { useUpdateStore } from '@/stores/appSettingsStore';
+  import { useUpdateStore } from '@/stores/updateStore';
   import { Button } from '@/components/ui/button';
   import { storeToRefs } from 'pinia';
 
@@ -25,14 +25,25 @@
       <DialogDescription>
         {{ `v${updateInfo?.version} 버전으로 업데이트 할 수 있습니다.` }}
       </DialogDescription>
+
+      <div
+        v-if="updateInfo?.releaseNotes"
+        class="text-muted-foreground max-h-32 overflow-y-auto rounded-md border p-3 text-sm"
+        v-html="updateInfo.releaseNotes"
+      />
+
       <div v-if="isDownloading">
         <p>{{ `다운로드 중... (${Math.round(downloadProgress)}%)` }}</p>
       </div>
 
       <DialogFooter class="gap-2">
-        <Button variant="ghost" @click="skipUpdate">{{ '이번 버전은 건너뛰기' }}</Button>
-        <Button variant="outline" @click="downloadLater">{{ '다음 시작 시 적용' }}</Button>
-        <Button @click="downloadNow">{{ '지금 업데이트' }}</Button>
+        <Button variant="ghost" :disabled="isDownloading" @click="skipUpdate">{{
+          '이번 버전은 건너뛰기'
+        }}</Button>
+        <Button variant="outline" :disabled="isDownloading" @click="downloadLater">{{
+          '다음 시작 시 적용'
+        }}</Button>
+        <Button :disabled="isDownloading" @click="downloadNow">{{ '지금 업데이트' }}</Button>
       </DialogFooter>
     </DialogContent>
   </Dialog>
